// src/component/_generated/api.ts
/* eslint-disable */
/**
 * Generated API types for the component.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type { AnyApi } from "convex/server";

/**
 * A type describing the component's public API.
 */
export const api: {
  public: {
    isMember: any;
    hasAccess: any;
    getAllGroups: any;
    getAllMembers: any;
    addMember: any;
    removeMember: any;
    grantAccess: any;
    revokeAccess: any;
    makeInvite: any;
    revokeInvite: any;
    revokeInviteByToken: any;
    listInvites: any;
    deleteOldInvites: any;
    claimInvite: any;
    hasInvite: any;
    getInviteByToken: any;
    upsertGroup: any;
    getGroup: any;
  };
} = null as any;

export const internal: any = null as any;